+++ src/main/java/org/rciam/plugins/groups/services/GroupAdminGroupMembers.java
@@ -72,42 +72,38 @@ public class GroupAdminGroupMembers {
     @Consumes("application/json")
     public Response inviteUser(GroupInvitationInitialRepresentation groupInvitationInitialRep) {
 
-        if (!isGroupAdmin){
+        if (!isGroupAdmin) {
             throw new ForbiddenException();
         }
 
         if (groupInvitationInitialRep.getEmail() == null || (groupInvitationInitialRep.isWithoutAcceptance() && groupInvitationInitialRep.getGroupEnrollmentConfiguration() == null))
-            throw new ErrorResponseException("Wrong data","Wrong data", Response.Status.BAD_REQUEST);
+            throw new ErrorResponseException("Wrong data", "Wrong data", Response.Status.BAD_REQUEST);
 
-        String emailId = group.getId();
-        Long invitationExpirationHour = null;
-        if (groupInvitationInitialRep.isWithoutAcceptance()) {
-            GroupEnrollmentConfigurationEntity conf = groupEnrollmentConfigurationRepository.getEntity(groupInvitationInitialRep.getGroupEnrollmentConfiguration().getId());
-            if (conf == null)
-                throw new ErrorResponseException("Wrong group enrollment configuration", "Wrong group enrollment configuration", Response.Status.BAD_REQUEST);
-            emailId = groupInvitationRepository.createForMember(groupInvitationInitialRep, groupAdmin.getId(),conf);
-            //execute once delete invitation after "url-expiration-period" ( default 72 hours)
-            AgmTimerProvider timer = session.getProvider(AgmTimerProvider.class);
-            invitationExpirationHour = realm.getAttribute(Utils.invitationExpirationPeriod) != null ? Long.valueOf(realm.getAttribute(Utils.invitationExpirationPeriod)) : 72;
-            long interval = invitationExpirationHour * 3600 * 1000;
-            timer.scheduleOnce(new ClusterAwareScheduledTaskRunner(session.getKeycloakSessionFactory(), new DeleteExpiredInvitationTask(emailId, realm.getId()), interval), interval, "DeleteExpiredInvitation_"+emailId);
-        }
+        GroupEnrollmentConfigurationEntity conf = groupEnrollmentConfigurationRepository.getEntity(groupInvitationInitialRep.getGroupEnrollmentConfiguration().getId());
+        if (conf == null)
+            throw new ErrorResponseException("Wrong group enrollment configuration", "Wrong group enrollment configuration", Response.Status.BAD_REQUEST);
+        String emailId = groupInvitationRepository.createForMember(groupInvitationInitialRep, groupAdmin.getId(), conf);
+        //execute once delete invitation after "url-expiration-period" ( default 72 hours)
+        AgmTimerProvider timer = session.getProvider(AgmTimerProvider.class);
+        Long invitationExpirationHour = realm.getAttribute(Utils.invitationExpirationPeriod) != null ? Long.valueOf(realm.getAttribute(Utils.invitationExpirationPeriod)) : 72;
+        long interval = invitationExpirationHour * 3600 * 1000;
+        timer.scheduleOnce(new ClusterAwareScheduledTaskRunner(session.getKeycloakSessionFactory(), new DeleteExpiredInvitationTask(emailId, realm.getId()), interval), interval, "DeleteExpiredInvitation_" + emailId);
+
 
         try {
             UserAdapter user = Utils.getDummyUser(groupInvitationInitialRep.getEmail(), groupInvitationInitialRep.getFirstName(), groupInvitationInitialRep.getLastName());
             customFreeMarkerEmailTemplateProvider.setUser(user);
-            customFreeMarkerEmailTemplateProvider.sendGroupInvitationEmail(groupAdmin, group.getName(), ModelToRepresentation.buildGroupPath(group), group.getFirstAttribute(Utils.DESCRIPTION), groupInvitationInitialRep.isWithoutAcceptance(), groupInvitationInitialRep.getGroupRoles(), emailId, invitationExpirationHour);
+            customFreeMarkerEmailTemplateProvider.sendGroupInvitationEmail(groupAdmin, group.getName(), ModelToRepresentation.buildGroupPath(group), group.getFirstAttribute(Utils.DESCRIPTION), groupInvitationInitialRep.getGroupRoles(), emailId, invitationExpirationHour);
 
-            if (groupInvitationInitialRep.isWithoutAcceptance()) {
-                groupAdminRepository.getAllAdminIdsGroupUsers(group).filter(x -> !groupAdmin.getId().equals(x)).map(id -> session.users().getUserById(realm, id)).forEach(admin -> {
-                    try {
-                        customFreeMarkerEmailTemplateProvider.setUser(admin);
-                        customFreeMarkerEmailTemplateProvider.sendInvitionAdminInformationEmail(user.getEmail(), true, group.getName(), groupAdmin, groupInvitationInitialRep.getGroupRoles());
-                    } catch (EmailException e) {
-                        throw new RuntimeException(e);
-                    }
-                });
-            }
+            groupAdminRepository.getAllAdminIdsGroupUsers(group).filter(x -> !groupAdmin.getId().equals(x)).map(id -> session.users().getUserById(realm, id)).forEach(admin -> {
+                try {
+                    customFreeMarkerEmailTemplateProvider.setUser(admin);
+                    customFreeMarkerEmailTemplateProvider.sendInvitionAdminInformationEmail(user.getEmail(), true, group.getName(), groupAdmin, groupInvitationInitialRep.getGroupRoles());
+                } catch (EmailException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+
         } catch (EmailException e) {
             ServicesLogger.LOGGER.failedToSendEmail(e);
         }